+++ internal/model/dto/response/queue.go
@@ -3,18 +3,18 @@ package response
 import "time"
 
 type QueueJobResponse struct {
-	JobID       uint      `json:"job_id"`
+	JobID       int       `json:"job_id"`
 	UserName    string    `json:"user_name"`
 	JobName     string    `json:"job_name"`
 	Description string    `json:"description"`
 	Status      int       `json:"status"`
 	SubmittedAt time.Time `json:"submitted_at"`
-	WaitSeconds int64     `json:"wait_seconds"`
+	WaitSeconds int       `json:"wait_seconds"`
 	FrontCount  int       `json:"front_count"`
 }
 
 type QueueJobDBRow struct {
-	JobID       uint
+	JobID       int
 	JobName     string
 	Description string
 	Status      int
