+++ internal/model/entity/queue.go
@@ -2,6 +2,6 @@ package entity
 
 // Item 队列中的元素（Scheduler 用）
 type Item struct {
-	JobID uint
-	Score int64
+	JobID int
+	Score int
 }
