+++ pkg/response/response.go
@@ -95,14 +95,14 @@ type PageRequest struct {
 // PageResponse 分页响应结构
 type PageResponse struct {
 	List     interface{} `json:"list"`     // 数据列表
-	Total    int64       `json:"total"`    // 总记录数
+	Total    int         `json:"total"`    // 总记录数
 	Page     int         `json:"page"`     // 当前页码
 	PageSize int         `json:"pageSize"` // 每页大小
 	Pages    int         `json:"pages"`    // 总页数
 }
 
 // NewPageResponse 创建分页响应
-func NewPageResponse(list interface{}, total int64, page, pageSize int) *PageResponse {
+func NewPageResponse(list interface{}, total int, page, pageSize int) *PageResponse {
 	pages := int(math.Ceil(float64(total) / float64(pageSize)))
 	return &PageResponse{
 		List:     list,
